services/notebook-service/requirements.txt
fastapi
uvicorn[standard]
nbformat
python-pptx
kaleido
pydantic
pytest
httpx
python-multipart
